Use 64bit frame Seq to prevent nonce reuse

+++ internal/multiplex/frame.go
@@ -2,7 +2,7 @@ package multiplex
 
 type Frame struct {
 	StreamID uint32
-	Seq      uint32
+	Seq      uint64
 	Closing  uint8
 	Payload  []byte
 }
